Split SQ8↔FP16 F16C kernels into sibling TUs [MOD-14954]
The SQ8↔FP16 kernels in the SSE4, AVX2, and AVX2+FMA tiers depend on
F16C (`_mm_cvtph_ps` / `_mm256_cvtph_ps`), while every other kernel in
those dispatcher TUs is F16C-clean. The previous arrangement mixed
both under `#ifdef OPT_F16C` blocks inside the base dispatcher .cpp/.h
files.

Split each tier's F16C-dependent kernels off into a sibling TU:

  functions/SSE4.cpp           → SSE4 + SQ8_FP32 (no F16C)
  functions/SSE4_F16C.cpp      → SQ8_FP16 only (requires -mavx -mf16c)

  functions/AVX2.cpp           → AVX2 + BF16 + SQ8_FP32 (no F16C)
  functions/AVX2_F16C.cpp      → SQ8_FP16 only (requires -mf16c)

  functions/AVX2_FMA.cpp       → SQ8_FP32 (no F16C)
  functions/AVX2_FMA_F16C.cpp  → SQ8_FP16 only (requires -mf16c)

The AVX-512 tier is unaffected — its SQ8_FP16 kernel uses
`_mm512_cvtph_ps`, which is part of AVX-512F and not F16C.

CMake now compiles each sibling TU conditionally on
`_has_full_f16c` and applies the F16C flags only there. Base TUs no
longer carry `-mf16c`, since they no longer reference F16C intrinsics.

Result:
- No `#ifdef OPT_F16C` directives in `functions/*.cpp` or `functions/*.h`.
- Compile-time isolation: an F16C intrinsic accidentally added outside
  a `_F16C` sibling will fail to build, not silently miscompile.
- Caller sites (`IP_space.cpp`, `L2_space.cpp`, `test_spaces.cpp`,
  `bm_spaces.h`) still gate the *calls* with `#ifdef OPT_F16C`; the new
  sibling .h includes are unconditional, since declarations alone don't
  link-error and the calls remain guarded.

Verified: 131 SQ8_FP16 unit tests + 115 ASan + 1166 full test_spaces
suite (covers other ISA tiers SQ8_FP32 / BF16 / INT8 / UINT8 to confirm
no regression from the dispatcher restructure).
