+++ src/VecSim/spaces/CMakeLists.txt
@@ -57,28 +57,31 @@ if(CMAKE_SYSTEM_PROCESSOR MATCHES "(x86_64)|(AMD64|amd64)|(^i.86$)")
 		set(_has_full_f16c TRUE)
 	endif()
 
+	# Base AVX2 / AVX2+FMA dispatcher TUs hold only kernels with no F16C dependency.
+	# SQ8↔FP16 kernels (which require F16C) live in sibling TUs functions/AVX2_F16C.cpp and
+	# functions/AVX2_FMA_F16C.cpp, compiled only when _has_full_f16c is true.
 	if(CXX_AVX2)
-		set(_avx2_flags "-mavx2")
-		if(_has_full_f16c)
-			message("Building functions/AVX2.cpp with AVX2 and F16C")
-			set(_avx2_flags "${_avx2_flags} -mf16c")
-		else()
-			message("Building functions/AVX2.cpp with AVX2 (no F16C)")
-		endif()
-		set_source_files_properties(functions/AVX2.cpp PROPERTIES COMPILE_FLAGS "${_avx2_flags}")
+		message("Building functions/AVX2.cpp with AVX2")
+		set_source_files_properties(functions/AVX2.cpp PROPERTIES COMPILE_FLAGS "-mavx2")
 		list(APPEND OPTIMIZATIONS functions/AVX2.cpp)
 	endif()
 
+	if(CXX_AVX2 AND _has_full_f16c)
+		message("Building functions/AVX2_F16C.cpp with AVX2 and F16C")
+		set_source_files_properties(functions/AVX2_F16C.cpp PROPERTIES COMPILE_FLAGS "-mavx2 -mf16c")
+		list(APPEND OPTIMIZATIONS functions/AVX2_F16C.cpp)
+	endif()
+
 	if(CXX_AVX2 AND CXX_FMA)
-		set(_avx2_fma_flags "-mavx2 -mfma")
-		if(_has_full_f16c)
-			message("Building functions/AVX2_FMA.cpp with AVX2, FMA, and F16C")
-			set(_avx2_fma_flags "${_avx2_fma_flags} -mf16c")
-		else()
-			message("Building functions/AVX2_FMA.cpp with AVX2 and FMA (no F16C)")
-		endif()
-		set_source_files_properties(functions/AVX2_FMA.cpp PROPERTIES COMPILE_FLAGS "${_avx2_fma_flags}")
+		message("Building functions/AVX2_FMA.cpp with AVX2 and FMA")
+		set_source_files_properties(functions/AVX2_FMA.cpp PROPERTIES COMPILE_FLAGS "-mavx2 -mfma")
 		list(APPEND OPTIMIZATIONS functions/AVX2_FMA.cpp)
+	endif()
+
+	if(CXX_AVX2 AND CXX_FMA AND _has_full_f16c)
+		message("Building functions/AVX2_FMA_F16C.cpp with AVX2, FMA, and F16C")
+		set_source_files_properties(functions/AVX2_FMA_F16C.cpp PROPERTIES COMPILE_FLAGS "-mavx2 -mfma -mf16c")
+		list(APPEND OPTIMIZATIONS functions/AVX2_FMA_F16C.cpp)
 	endif()
 
 	if(CXX_F16C AND CXX_FMA AND CXX_AVX)
@@ -100,17 +103,17 @@ if(CMAKE_SYSTEM_PROCESSOR MATCHES "(x86_64)|(AMD64|amd64)|(^i.86$)")
 	endif()
 
 	if(CXX_SSE4)
-		set(_sse4_flags "-msse4.1")
-		if(_has_full_f16c)
-			# F16C is VEX-encoded → must compile with -mavx alongside -mf16c, matching the
-			# F16C.cpp recipe above.
-			message("Building functions/SSE4.cpp with SSE4.1, AVX, and F16C")
-			set(_sse4_flags "${_sse4_flags} -mavx -mf16c")
-		else()
-			message("Building functions/SSE4.cpp with SSE4.1 (no F16C)")
-		endif()
-		set_source_files_properties(functions/SSE4.cpp PROPERTIES COMPILE_FLAGS "${_sse4_flags}")
+		message("Building functions/SSE4.cpp with SSE4.1")
+		set_source_files_properties(functions/SSE4.cpp PROPERTIES COMPILE_FLAGS "-msse4.1")
 		list(APPEND OPTIMIZATIONS functions/SSE4.cpp)
+	endif()
+
+	# SSE4 SQ8↔FP16 kernels need F16C, which is VEX-encoded → require -mavx alongside -mf16c
+	# (mirrors the F16C.cpp recipe above).
+	if(CXX_SSE4 AND _has_full_f16c)
+		message("Building functions/SSE4_F16C.cpp with SSE4.1, AVX, and F16C")
+		set_source_files_properties(functions/SSE4_F16C.cpp PROPERTIES COMPILE_FLAGS "-msse4.1 -mavx -mf16c")
+		list(APPEND OPTIMIZATIONS functions/SSE4_F16C.cpp)
 	endif()
 
 	if(CXX_SSE)
