+++ src/VecSim/spaces/functions/AVX2.h
@@ -19,10 +19,5 @@ dist_func_t<float> Choose_SQ8_FP32_L2_implementation_AVX2(size_t dim);
 dist_func_t<float> Choose_BF16_IP_implementation_AVX2(size_t dim);
 dist_func_t<float> Choose_BF16_L2_implementation_AVX2(size_t dim);
 
-#ifdef OPT_F16C
-dist_func_t<float> Choose_SQ8_FP16_IP_implementation_AVX2(size_t dim);
-dist_func_t<float> Choose_SQ8_FP16_Cosine_implementation_AVX2(size_t dim);
-dist_func_t<float> Choose_SQ8_FP16_L2_implementation_AVX2(size_t dim);
-#endif
 
 } // namespace spaces
